# backend/swagger/doc/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Standard Api Server",
        "title": "Prototype Api Server",
        "contact": {
            "name": "sigma",
            "url": "https://github.com/sigmalin/Prototype"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:80/api",
    "paths": {
...
        "/users/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "login token",
                        "name": "token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/login.Result"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Login Failure",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bankData.Content": {
            "type": "object",
            "properties": {
                "Coin": {
                    "type": "integer",
                    "example": 1000
                },
                "Faith": {
                    "type": "integer",
                    "example": 0
                },
                "Gems": {
                    "type": "integer",
                    "example": 0
                },
                "Treasure": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
...
        "login.Result": {
            "type": "object",
            "properties": {
                "Bank": {
                    "$ref": "#/definitions/bankData.Content"
                },
                "JsonWebToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                }
            }
        },
        "response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": ""
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}